Crate sui_framework_snapshot

Structs

Snapshot

Encapsulation of an entry in the manifest file corresponding to a single version of the system packages.

SnapshotPackage

Entry in the manifest file corresponding to a specific version of a specific system package

Functions

load_bytecode_snapshot

load_bytecode_snapshot_manifest

manifest_path

update_bytecode_snapshot_manifest

Type Aliases

SnapshotManifest